+++ httprunner/plugins/locusts/__main__.py
@@ -1,6 +1,4 @@
-import sys
-
-from httprunner.plugins.locusts import main
+from httprunner.plugins.locusts.cli import main
 
 if __name__ == "__main__":
     main()
